refactor: 커스텀 어노테이션 부분을 도메인 외부로 이전
